+++ R/read_redcap.R
@@ -53,7 +53,7 @@
 #' @param guess_max A positive [base::numeric] value
 #' passed to [readr::read_csv()] that specifies the maximum number of records to
 #' use for guessing column types. Default `.Machine$integer.max`.
-#' @param enable_mixed_structure A logical to allow for support of mixed repeating/non-repeating
+#' @param allow_mixed_structure A logical to allow for support of mixed repeating/non-repeating
 #' instruments. Setting to `TRUE` will treat the mixed instrument's non-repeating versions
 #' as repeating instruments with a single instance. Applies to longitudinal projects
 #' only. Default `FALSE`.
@@ -80,7 +80,7 @@ read_redcap <- function(redcap_uri,
                         export_data_access_groups = NULL,
                         suppress_redcapr_messages = TRUE,
                         guess_max = .Machine$integer.max,
-                        enable_mixed_structure = FALSE) {
+                        allow_mixed_structure = FALSE) {
   check_arg_is_character(redcap_uri, len = 1, any.missing = FALSE)
   check_arg_is_character(token, len = 1, any.missing = FALSE)
   check_arg_is_valid_token(token)
@@ -89,6 +89,7 @@ read_redcap <- function(redcap_uri,
   check_arg_is_logical(export_survey_fields, len = 1, any.missing = FALSE, null.ok = TRUE)
   check_arg_is_logical(export_data_access_groups, len = 1, any.missing = FALSE, null.ok = TRUE)
   check_arg_is_logical(suppress_redcapr_messages, len = 1, any.missing = FALSE)
+  check_arg_is_logical(allow_mixed_structure, len = 1, any.missing = FALSE)
 
   # Load REDCap Metadata ----
   # Capture unexpected metadata API call errors
@@ -273,7 +274,7 @@ read_redcap <- function(redcap_uri,
       db_data_long = db_data,
       db_metadata_long = db_metadata,
       linked_arms = linked_arms,
-      enable_mixed_structure = enable_mixed_structure
+      allow_mixed_structure = allow_mixed_structure
     )
   } else {
     out <- clean_redcap(